requirements.txt
discord.py
asyncio
pyperclip
firebase-admin